+++ baby-bear/src/x86_64_avx2/packing.rs
