+++ baby-bear/src/x86_64_avx512/mod.rs
@@ -1,4 +1,3 @@
-#[allow(clippy::module_inception)]
-mod x86_64_avx512;
+mod packing;
 
-pub use x86_64_avx512::*;
+pub use packing::*;
