+++ baby-bear/src/x86_64_avx512/packing.rs
