+++ mersenne-31/src/aarch64_neon/mod.rs
@@ -1,4 +1,3 @@
-#[allow(clippy::module_inception)]
-mod aarch64_neon;
+mod packing;
 
-pub use aarch64_neon::*;
+pub use packing::*;
