+++ mersenne-31/src/x86_64_avx2/mod.rs
@@ -1,4 +1,3 @@
-#[allow(clippy::module_inception)]
-mod x86_64_avx2;
+mod packing;
 
-pub use x86_64_avx2::*;
+pub use packing::*;
