+++ mersenne-31/src/x86_64_avx2/packing.rs
